refactor : #24 destination -> recommend 로 이름 수정
