+++ src/app/(provider)/(root)/recommend/_components/AreaCard.tsx
