+++ src/app/(provider)/(root)/recommend/_components/DetailCard.tsx
@@ -10,7 +10,7 @@ type DetailCardProps = {
 function DetailCard({ title, description, imageUrl }: DetailCardProps) {
   return (
     <>
-      <ImageContainer isTitle size="detail" imageUrl={imageUrl} />
+      <ImageContainer isTitle size="detail" imageUrl={imageUrl || "/"} />
       <CardForm title={title} description={description} intent="detail" />
     </>
   );
