+++ src/app/(provider)/(root)/recommend/area/[id]/page.tsx
